Mudança na persistência dos produtos, agora o ID é atualizado automaticamente com o do banco de dados. Os métodos salvar está retornando o objeto persistindo e não mais um boolean.

+++ Offstation/src/projetos/gerencia/negocio/produto/Peca.java
@@ -2,7 +2,7 @@ package projetos.gerencia.negocio.produto;
 
 public class Peca extends Produto {
 
-    public Peca(int id, int estoque, double preco, String nome, String marca) {
+    public Peca(long id, int estoque, double preco, String nome, String marca) {
         super(id, estoque, preco, nome, marca);
     }
