+++ Offstation/src/projetos/gerencia/negocio/produto/Servico.java
@@ -2,7 +2,7 @@ package projetos.gerencia.negocio.produto;
 
 public class Servico extends Produto {
 
-    public Servico(int id, int estoque, double preco, String nome, String marca) {
+    public Servico(long id, int estoque, double preco, String nome, String marca) {
         super(id, estoque, preco, nome, marca);
     }
